HW01 add form adding

+++ src/main/java/ru/javawebinar/topjava/web/MealServlet.java
@@ -3,6 +3,7 @@ package ru.javawebinar.topjava.web;
 import org.slf4j.Logger;
 import ru.javawebinar.topjava.dao.DAO;
 import ru.javawebinar.topjava.dao.DAO_implementation;
+import ru.javawebinar.topjava.util.MealsUtil;
 
 
 import javax.servlet.ServletException;
@@ -10,6 +11,7 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.time.LocalTime;
 
 
 import static org.slf4j.LoggerFactory.getLogger;
@@ -27,10 +29,13 @@ public class MealServlet extends HttpServlet{
         //delete by id
         String deleteByIdString = request.getParameter("deleteById");
         Integer deleteById = null;
-        if(deleteByIdString != null) {
+        if(deleteByIdString != null && deleteByIdString != "") {
             deleteById = Integer.parseInt(deleteByIdString);
             dao.deleteById(deleteById);
         }
+
+        //form trigger
+        request.setAttribute("formTrigger", true);
 
         //get id by onclick row action
         String idString = request.getParameter("id");
@@ -39,7 +44,7 @@ public class MealServlet extends HttpServlet{
         System.out.println(id);
         request.setAttribute("id", id);
 
-        request.setAttribute("mealsList", dao.read());
+        request.setAttribute("mealsList", MealsUtil.getFilteredWithExceeded(dao.read(), LocalTime.MIN,LocalTime.MAX,2000));
 
         request.getRequestDispatcher("/mealList.jsp").forward(request, response);
     }
